+++ ChangeLog.md
@@ -26,6 +26,7 @@ and this project adheres to
 - word wrap with function names in PDF cheatsheet
 - `shell/process-info` to handle Java Optional type better
 - callstack for macro arity exception
+- bindings in `let`, `recur`, and `bindings` to throw an exception if unbalanced